web-ui/backend/requirements.txt
# Web UI 后端依赖
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
